requirements.txt
requests==2.31.0
pymongo==4.6.1
python-dotenv==1.0.0
Pillow==10.2.0
pytest==7.4.4
pytest-mock==3.12.0
responses==0.24.1
